Fix scrobbling and play count increasing
There is no robust way for a backend to know when a track is done
playing. Thus I added a GMusicScrobblerFrontend that listens for the
track_playback_ended event, checks if we've played enough of the track,
and if so sends an event to the GMusicBackend, which uses gmusicapi to
increase the play count. Fixes #96.

By removing the play count management from GMusicPlaybackProvider, I
could strip it down to a pure and simple translate_uri(). As a side
effect of this, Last.fm scrobbling through Mopidy-Scrobbler now works.
Fixes #60. Fixes mopidy/mopidy-scrobbler#6.

As I could find no official policy for when Google Play Music's play
count should be increased, I changed it from requiring 2/3 of the track
to have been played to 50% or 240s, which aligns with Last.fm's rules
as used in Mopidy-Scrobbler.