Document refresh configuration
I was going to add a feature to allow the refresh intervals to be configurable, but looking at the code, I realised it already is, just undocumented.
I was finding that the playlist refresh every 60 minutes was causing the CPU to max out on my Pi Zero, causing it to be unusable for a couple of minutes. I mentioned that in case anyone else encounters the same issue.